Moved mutual meta code into Meta class
